requirements.txt
# These are the libraries required to run your app
# They DO NOT include libraries needed to develop or test our app
#
jinja2~=3.1.4
pyyaml~=6.0.2
